# Conflicts:
#	Integration/SlideSentinelBase/lib/Properties/Properties.h
#	Integration/SlideSentinelRover/.vscode/settings.json
#	Integration/SlideSentinelRover/lib/PropHandler/PropHandler.cpp
#	Integration/SlideSentinelRover/lib/PropHandler/PropHandler.h
#	Integration/SlideSentinelRover/lib/Properties/Properties.cpp
#	Integration/SlideSentinelRover/lib/RoverModel/SSModel.cpp
#	Integration/SlideSentinelRover/lib/RoverModel/SSModel.h
#	Integration/SlideSentinelRover/lib/SSModel/SSModel.cpp
#	Integration/SlideSentinelRover/lib/SSModel/SSModel.h
#	Integration/SlideSentinelRover/libhw/PropHandler/PropHandler.cpp
#	Integration/SlideSentinelRover/libhw/PropHandler/PropHandler.h
#	Integration/SlideSentinelRover/libhw/SSModel/SSModel.cpp
#	Integration/SlideSentinelRover/libhw/SSModel/SSModel.h
#	Integration/SlideSentinelRover/platformio.ini
#	Integration/SlideSentinelRover/src/config_2.0.0.h

+++ Integration/SlideSentinelRover/libhw/FSController/FSController.h
@@ -9,8 +9,8 @@
 #define MAIN "data"
 #define SD_SIZE 16 // GB
 
-// TODO consider deep copying over the timestamp so we dont depend on a string not
-// within the FSController class
+// TODO consider deep copying over the timestamp so we dont depend on a string
+// not within the FSController class
 class FSController : public Controller {
 private:
   SdFat m_sd;
@@ -24,8 +24,8 @@ private:
   char *m_curDir;
 
   // TODO add error count diagnostic info
-  float m_spaceMB;  // diagnostic
-  uint16_t m_cycle; // diagnostic
+  float m_spaceMB; // diagnostic
+  int m_cycle;     // diagnostic
   void m_cycles();
   void m_SDspace();
